+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
  - On x86:
    - Support for running on Xeon Phi processors.
    - Removed the `ucode=allow-same` command line option.
+   - Removed x2APIC Cluster Mode for external interrupts.  x2APIC Physical and
+     Mixed Modes are still available.
 
 ## [4.19.0](https://xenbits.xenproject.org/gitweb/?p=xen.git;a=shortlog;h=RELEASE-4.19.0) - 2024-07-29